fix information about versions
There is currently no 4.X.X branch which is confusing when trying
to follow the guide. The section was split into two subsections
about 3.X.X and 4.X.X. A note was added about the 4.X.X branch
missing (could be a callout but configuration would need to be
changed for that I think).

The same information was also present under `Kompilieren vorbereiten`
so it was removed there.